fix(valid-expect): look past parentheses when checking for await or return
