+++ internal/plugins/rstest/rules/valid_expect/valid_expect.go
@@ -10,7 +10,7 @@ import (
 	rstestUtils "github.com/web-infra-dev/rslint/internal/plugins/rstest/utils"
 	"github.com/web-infra-dev/rslint/internal/rule"
 	internalUtils "github.com/web-infra-dev/rslint/internal/utils"
-	testFramework "github.com/web-infra-dev/rslint/internal/utils/test_framework"
+	sharedValidExpect "github.com/web-infra-dev/rslint/internal/utils/test_framework/rules/valid_expect"
 )
 
 //go:embed valid_expect.schema.json
@@ -164,117 +164,9 @@ func readIntOption(options map[string]interface{}, key string, defaultValue int)
 // --- Async assertion machinery ---
 //
 // The Promise-chain walking, acceptable-return detection, array de-duplication
-// and async/await fixers below are ported verbatim from
-// internal/plugins/jest/rules/valid_expect/valid_expect.go, which is itself
-// aligned with eslint-plugin-vitest. Kept as a local copy rather than shared:
-// sharing them would require touching the jest rule and re-running its
-// regression, and there is no third consumer yet.
-
-func isPromiseMethodCall(node *ast.Node) bool {
-	if node == nil || node.Kind != ast.KindCallExpression {
-		return false
-	}
-
-	callee := ast.SkipParentheses(node.AsCallExpression().Expression)
-	if !rstestUtils.IsMemberAccessNode(callee) {
-		return false
-	}
-
-	return testFramework.CalleeChainName(internalUtils.AccessExpressionObject(callee)) == "Promise"
-}
-
-func getPromiseCallExpressionNode(node *ast.Node) *ast.Node {
-	if node == nil {
-		return nil
-	}
-
-	if node.Kind == ast.KindArrayLiteralExpression && node.Parent != nil && node.Parent.Kind == ast.KindCallExpression {
-		node = node.Parent
-	}
-
-	if isPromiseMethodCall(node) {
-		return node
-	}
-
-	return nil
-}
-
-func findPromiseCallExpressionNode(node *ast.Node) *ast.Node {
-	if node == nil || node.Parent == nil || node.Parent.Parent == nil {
-		return nil
-	}
-	if node.Parent.Kind != ast.KindCallExpression && node.Parent.Kind != ast.KindArrayLiteralExpression {
-		return nil
-	}
-	return getPromiseCallExpressionNode(node.Parent)
-}
-
-func getParentIfPromiseChained(node *ast.Node) *ast.Node {
-	if node == nil || node.Parent == nil || node.Parent.Parent == nil {
-		return node
-	}
-
-	grandParent := node.Parent.Parent
-	if grandParent.Kind != ast.KindCallExpression || !rstestUtils.IsMemberAccessNode(grandParent.AsCallExpression().Expression) {
-		return node
-	}
-
-	member := grandParent.AsCallExpression().Expression
-	entries := testFramework.GetMemberEntries(member)
-	if len(entries) == 0 {
-		return node
-	}
-
-	last := entries[len(entries)-1].Name
-	if last == "then" || last == "catch" {
-		return getParentIfPromiseChained(grandParent)
-	}
-
-	return node
-}
-
-func isAcceptableReturnNode(node *ast.Node, allowReturn bool) bool {
-	if node == nil {
-		return false
-	}
-
-	if allowReturn && node.Kind == ast.KindReturnStatement {
-		return true
-	}
-	if node.Kind == ast.KindConditionalExpression {
-		return isAcceptableReturnNode(node.Parent, allowReturn)
-	}
-
-	return node.Kind == ast.KindArrowFunction || node.Kind == ast.KindAwaitExpression
-}
-
-func promiseArrayExceptionKey(sourceFile *ast.SourceFile, node *ast.Node) string {
-	if sourceFile == nil || node == nil {
-		return ""
-	}
-	r := internalUtils.TrimNodeTextRange(sourceFile, node)
-	return fmt.Sprintf("%d:%d", r.Pos(), r.End())
-}
-
-func asyncInsertFix(sourceFile *ast.SourceFile, fn *ast.Node) rule.RuleFix {
-	switch fn.Kind {
-	case ast.KindMethodDeclaration, ast.KindConstructor, ast.KindGetAccessor, ast.KindSetAccessor:
-		head := internalUtils.GetFunctionHeadLoc(sourceFile, fn)
-		return rule.RuleFixReplaceRange(core.NewTextRange(head.Pos(), head.Pos()), "async ")
-	default:
-		return rule.RuleFixInsertBefore(sourceFile, fn, "async ")
-	}
-}
-
-func awaitFix(sourceFile *ast.SourceFile, node *ast.Node, alwaysAwait bool) rule.RuleFix {
-	if alwaysAwait && node.Parent != nil && node.Parent.Kind == ast.KindReturnStatement {
-		ret := node.Parent
-		retRange := internalUtils.TrimNodeTextRange(sourceFile, ret)
-		nodeRange := internalUtils.TrimNodeTextRange(sourceFile, node)
-		return rule.RuleFixReplaceRange(core.NewTextRange(retRange.Pos(), nodeRange.Pos()), "await ")
-	}
-	return rule.RuleFixInsertBefore(sourceFile, node, "await ")
-}
+// and async/await fixers live in the shared package: they are pure syntax, with
+// no jest or rstest semantics, and keeping two copies is what let the same
+// parenthesized-assertion false positive exist twice.
 
 func reportAsyncDescriptor(
 	ctx rule.RuleContext,
@@ -287,10 +179,10 @@ func reportAsyncDescriptor(
 	var fixes []rule.RuleFix
 	if fn := ast.GetContainingFunction(descriptor.node); fn != nil {
 		if !ast.IsAsyncFunction(fn) && !asyncInserted[fn] {
-			fixes = append(fixes, asyncInsertFix(ctx.SourceFile, fn))
+			fixes = append(fixes, sharedValidExpect.AsyncInsertFix(ctx.SourceFile, fn))
 			asyncInserted[fn] = true
 		}
-		fixes = append(fixes, awaitFix(ctx.SourceFile, descriptor.node, alwaysAwait))
+		fixes = append(fixes, sharedValidExpect.AwaitFix(ctx.SourceFile, descriptor.node, alwaysAwait))
 	}
 
 	if len(fixes) > 0 {
@@ -350,37 +242,6 @@ func isAllowedExtraExpectArg(arg *ast.Node, entry rstestUtils.RstestExpectEntry)
 
 func shouldBeAwaited(parsed *rstestUtils.ParsedRstestExpectCall, asyncMatchers []string) bool {
 	return rstestUtils.ShouldRstestExpectBeAwaited(parsed, asyncMatchers)
-}
-
-// resolveAsyncAssertionReportNode mirrors the jest rule: from the complete
-// assertion expression it walks out through chained .then/.catch, notes whether
-// the assertion sits inside a Promise.all([...]) array, and decides whether the
-// resulting node must be awaited or returned. assertionNode must be the
-// outermost expression of the chain — ParsedRstestExpectCall.Expression — so
-// that a Chai chain carrying several matchers, such as
-// expect(p).resolves.to.be.a("string").that.contains("x"), is inspected as a
-// whole rather than at its first matcher.
-func resolveAsyncAssertionReportNode(
-	assertionNode *ast.Node,
-	alwaysAwait bool,
-) (reportNode *ast.Node, promiseWrapped bool, insideAssertionArray bool, shouldReport bool) {
-	if assertionNode == nil {
-		return nil, false, false, false
-	}
-
-	promiseChainedAssertionNode := getParentIfPromiseChained(assertionNode)
-	insideAssertionArray = promiseChainedAssertionNode.Parent != nil && promiseChainedAssertionNode.Parent.Kind == ast.KindArrayLiteralExpression
-	reportNode = promiseChainedAssertionNode
-	if promiseCallNode := findPromiseCallExpressionNode(promiseChainedAssertionNode); promiseCallNode != nil {
-		reportNode = promiseCallNode
-		promiseWrapped = true
-	}
-
-	if reportNode.Parent == nil || isAcceptableReturnNode(reportNode.Parent, !alwaysAwait) {
-		return reportNode, promiseWrapped, insideAssertionArray, false
-	}
-
-	return reportNode, promiseWrapped, insideAssertionArray, true
 }
 
 // reportBrokenChain reports a chain that the parser resolved without a matcher.
@@ -457,7 +318,7 @@ var ValidExpectRule = rule.Rule{
 					return
 				}
 
-				reportNode, promiseWrapped, insideAssertionArray, shouldReport := resolveAsyncAssertionReportNode(
+				reportNode, promiseWrapped, insideAssertionArray, shouldReport := sharedValidExpect.ResolveAsyncAssertionReportNode(
 					parsed.Expression,
 					opts.AlwaysAwait,
 				)
@@ -465,7 +326,7 @@ var ValidExpectRule = rule.Rule{
 					return
 				}
 
-				reportNodeKey := promiseArrayExceptionKey(ctx.SourceFile, reportNode)
+				reportNodeKey := sharedValidExpect.PromiseArrayExceptionKey(ctx.SourceFile, reportNode)
 				if shouldReport && !arrayExceptions[reportNodeKey] {
 					descriptors = append(descriptors, asyncDescriptor{
 						node:           reportNode,
